processing_modelling/src/requirements.txt
# Core libraries
pandas==2.3.3
numpy==2.2.4

# Machine Learning
scikit-learn==1.7.2
xgboost==3.1.3

# String matching
python-Levenshtein==0.27.3
Unidecode==1.4.0

